+++ Professions/Framework/Patchers/Farming/FarmAnimalGetSellPricePatcher.cs
@@ -2,7 +2,6 @@
 
 #region using directives
 
-using System.Reflection;
 using DaLion.Shared.Harmony;
 using HarmonyLib;
 
@@ -23,24 +22,13 @@ internal sealed class FarmAnimalGetSellPricePatcher : HarmonyPatcher
     #region harmony patches
 
     /// <summary>Patch to adjust Breeder animal sell price.</summary>
-    [HarmonyPrefix]
+    [HarmonyPostfix]
     [UsedImplicitly]
-    private static bool FarmAnimalGetSellPricePrefix(FarmAnimal __instance, ref int __result)
+    private static void FarmAnimalGetSellPricePostfix(FarmAnimal __instance, ref int __result)
     {
-        if (!__instance.DoesOwnerHaveProfessionOrLax(Profession.Breeder))
+        if (__instance.DoesOwnerHaveProfessionOrLax(Profession.Breeder))
         {
-            return true; // run original logic
-        }
-
-        try
-        {
-            __result = (int)(__instance.getSellPrice() * __instance.GetBreederAdjustedPrice());
-            return false; // don't run original logic
-        }
-        catch (Exception ex)
-        {
-            Log.E($"Failed in {MethodBase.GetCurrentMethod()?.Name}:\n{ex}");
-            return true; // default to original logic
+            __result = (int)(__result * __instance.GetBreederAdjustedPrice());
         }
     }